requirements.txt
bluepy==1.1.0
requests==2.18.1
twisted==17.9.0
treq==17.8.0

w1thermsensor==1.0.5
paho-mqtt==1.3.1
python-dotenv==0.8.1
Adafruit-BME280==1.0.1